renamed a few components for consistency
