+++ src/components/layout/detailPages/detailPages.module.css
