--- src/components/member/memberList.js
@@ -1,11 +1,0 @@
-import MemberCard from "./memberCard";
-
-function MemberList() {
-  return (
-    <div>
-      <MemberCard />
-    </div>
-  );
-}
-
-export default MemberList;
